Page numbers from uint16 to uint32
increased the size for the page numbers from 16bit to 32bit, because,
what the heck, it's only 4 additional bytes per page, and this moves us
closer to infinity. The number of bits printed actually stayes the same,
only the page header loses some redundancy. The payload size of 64bit
does not really make sense for 64ki max pages anyway. If anyone wants to
print out and store 1 million pages of a file, who are we to stop them?